tests: Add a safe prepend function for path-like environment variables
Copy the prepend function from libnbd/run.in (written by me and
relicensed to BSD for this purpose).  Use it in a few places in the
tests.

This is refactoring, but in some cases it will change the final value
of the path, hopefully making it more correct than before.