tests/test-shebang-crlf.sh: Fix location of shebang-crlf.py
Commit 5d01b47d7a ("tests: Add a regression test for DOS line endings
in shebang plugins") requires a test file shebang-crlf.py which is
generated in the Makefile.  Generating this file was a late addition
to the commit, but this late change broke srcdir != builddir builds,
since the file is now in the builddir not the srcdir.

The error seen was:

../../tests/test-shebang-crlf.sh: could not locate /builddir/build/BUILD/nbdkit-1.39.4/build_native/../tests/shebang-crlf.py or not executable

Fixes: commit 5d01b47d7a2b1fc92fa516c074bd4deadf1a1e7a
(cherry picked from commit aade881287fbf5c7ef3539d2f01e0d6eea94a8a9)