common/utils: Add a function to find the size of a file or block device
It turns out our existing method was not reliable, at least on some
BSDs.  BSD treats SEEK_END as seeking to stat(2).st_size, but st_size
is 0 for block devices, thus always seeking to the start.

This method is copied and adapted from
https://github.com/util-linux/util-linux/blob/master/lib/blkdev.c

There is no version of this function for Windows.  We could add that,
as we have the code (plugins/file/winfile.c) but it requires quite a
different prototype.  At the moment it's best if code which tries to
use this function and is compiled on Windows fails hard so we can
identify those places and fix them.

Thanks: Eric Blake
(cherry picked from commit 910db9301faa3b85fed987e436730288515182f6)