tests: More debug output in test-blocksize-sharding.sh
We got a report of a test-blocksize-sharding.sh failure on ArchLinux
[1], with a log pointing to the sanity check taking longer than
expected.  Since there were insufficient breadcrumbs to make it easy
to see what went wrong, enhance the test to track more points in time
and avoid confusing output ordering.  Also, give the test a bit more
leeway (3.5 seconds for the sanity check is still okay).

[1] https://0x0.st/X8TI.log

CC: solstice.dhiver@gmail.com
Signed-off-by: Eric Blake <eblake@redhat.com>