server: Make the --print-uri option easier to machine parse
It seems better if the first line of the output is just the URI
without any extra fluff.  This makes it easier to machine parse this
output (although usually that is not necessary as you can just as
easily form the URI yourself).

Updates: commit fcf5073d1adb1e92e02cbfca9c1ca3d93c977dfa