server/uri.c: Append tls-certificates or tls-psk-key parameter to $uri
If TLS is required (--tls=require), and either --run or --print-uri is
used, include extra query parameters ?tls-certificates=... or
?tls-psk-key=... in the URI.

These are simply copied from the corresponding --tls-certificates or
--tls-psk parameter on the nbdkit command line without any other
modification (we don't make it an absolute path as it's not needed).

This will only work for libnbd-based clients.  qemu (currently) does
not support TLS NBD URIs at all, and requires the separate -object
parameter to pass certificates and PSK files.

This commit as it stands doesn't actually work well for PSK, unless
the username in the PSK file happens to coincide with the local login
name.

Example:

  $ nbdkit null \
           --tls=require --tls-certificates=tests/pki --tls-verify-peer \
           --print-uri \
           --run 'echo uri="$uri" ; nbdinfo "$uri"'
  nbds+unix://?socket=/tmp/nbdkitIdKUcE/socket&tls-certificates=tests/pki
  Shell-quoted URI: "nbds+unix://?socket=/tmp/nbdkitIdKUcE/socket&tls-certificates=tests/pki"
  Command to query the NBD endpoint:
    nbdinfo "nbds+unix://?socket=/tmp/nbdkitIdKUcE/socket&tls-certificates=tests/pki"
  uri=nbds+unix://?socket=/tmp/nbdkitIdKUcE/socket&tls-certificates=tests/pki
  protocol: newstyle-fixed with TLS, using structured packets
  export="":
  	export-size: 0
  	content: empty
  	uri: nbds+unix:///?socket=/tmp/nbdkitIdKUcE/socket&tls-certificates=tests/pki
  	contexts:
  		base:allocation
  	is_rotational: false
  	is_read_only: false
  	can_block_status_payload: false
  	can_cache: true
  	can_df: true
  	can_fast_zero: true
  	can_flush: true
  	can_fua: true
  	can_multi_conn: true
  	can_trim: true
  	can_zero: true