tests: Require bzip2 on OpenSUSE
OpenSUSE has an odd dependency issue: libbz2-devel does not pull in
the gzip2 command line tool.  Work around that in two ways: update the
CI definition to explicitly require the command line tool, and tweak
the makefile to avoid building the bzip2 filter when it can't be
tested.  (Technically, we could introduce two levels of conditionals:
one to build the filter when the libs are present and another to test
it when the bzip2 tool is present, but that feels like overkill)

Signed-off-by: Eric Blake <eblake@redhat.com>