file: Return minimum and preferred I/O size for block devices
When this plugin is serving a block device, read the minimum and
preferred I/O sizes and return those to the NBD client.  There is no
upper limit.

For files this does nothing at the moment.  In theory we could
synthesize that information as well from filesystem block size.

This document was useful:
https://access.redhat.com/articles/3911611