file: Rethink calculation of block device I/O size preferences
While testing this and after examining the kernel source I found that
BLKIOOPT is almost always returned as 0.  Only the MD (RAID) device
calculates it (as the stripe size).  Change the calculation to use max
of the minimum size and 4K in this case.

This commit also adds tests.  Testing the block size constraints
through libnbd works as expected.  However through libguestfs it does
not give the right results.  I didn't investigate this further as the
problem could be anywhere in qemu or Linux.

Also, fix error checking of ioctl to be consistent with other code in
the same source file.

Fixes: commit 525b242caae89f9b8fcfc9aba749516e4fd98703