tests: bzip2: Don't keep recreating disk.bz2
Unlike other compression programs it seems 'bzip2 --keep' tries to
preserve the original timestamp of the input file for the output
(.bz2) file.  Unfortunately this means it appears to 'make' that the
output file has not been updated, so we kept running 'bzip2' each time
we ran 'make check'.

The simple fix is to touch the output file.

Fixes: commit 85d504ac64679b591e1c6093089992fa42421c16