server: Take a thread-local copy of the last call to nbdkit_error
nbdkit_error has traditionally been a "fancy wrapper around fprintf"
(kind of, don't take that literally).  It is encouraged that plugins
and filters do something like:

  if (error) {
    nbdkit_error ("oops, a bad thing happened");
    return -1;
  }

but we don't enforce this.  Plugins might call nbdkit_error more than
once or not at all.

The point where we get to sending an error back over the wire to the
NBD client is long after the plugin returned above, and after
nbdkit_error was called.

Therefore in order to send errors back to the NBD client, we must keep
the last error message around.

This change simply modifies nbdkit_error to make a best-effort attempt
to save the last error message in thread-local storage.

We also clear the last error when a new request starts, to ensure that
we don't leak errors across different callbacks or connections.