server: Send the last error to the NBD client
This sends the last error saved in the connection handle back to the
NBD client.  This is informational and best effort.

qemu reports the error already, for example:

  $ nbdkit --log=null \
           eval open=' echo EPERM Go Away >&2; exit 1 ' get_size=' echo 100 ' \
           --run 'qemu-img info "$uri"'
  qemu-img: Could not open 'nbd+unix://?socket=/tmp/nbdkitIDl6iy/socket': Requested export not available
  server reported: /tmp/nbdkitRDAfXH/open: Go Away

This goes back to at least qemu 2.12.0 (RHEL 7) and possibly earlier,
so we can just assume that qemu does this for the test.

libnbd requires a patch to display this information.