tests/test-ext2-exportname.sh: Ignore libnbd < 1.8 for this test
This test attempted to detect a case that happens in libnbd 1.6.2
where an error was reported but the nbdinfo tool still exited with
EXIT_SUCCESS.  However this test was wrong and a recent change to
libnbd caused it to fail visibly.

The test was wrong because it checked the return code from cat ($?)
rather than the saved return code from nbdinfo ($st).

libnbd commit 474a4ae6c8 ("lib: Don't overwrite error in
nbd_opt_{go,info}") changed the error message so "server replied with
error" no longer appears, which visibly broke this test.

I chose to get rid of the incorrect test case and just skip this
entire test on libnbd < 1.8.  (This skips the test on RHEL 8.)  libnbd
1.8 was released over 3 years ago.