vddk: Detect possible VDDK crash and warn
VDDK >= 8 has a bug where if /sys/class/scsi_disk is missing, it will
crash when trying to scan for SAN mode.  Detect what we think is the
possible combination of causes and emit a debug message warning about
this and suggesting that the user read the bug.

Unfortunately we have to do this as a debug message (not an error),
and we have to do it after forking, because we must wait until VDDK
has been loaded to get the VDDK version.  Still, this is better than
nothing.

Thanks: Matthew Arnold, Nijin Ashok, Eric Blake
Related: https://issues.redhat.com/browse/RHEL-54377
(cherry picked from commit 6f399662b66edb4488bde41cf54f61907d8f3290)