vddk: Check create-size is aligned to VIXDISKLIB_SECTOR_SIZE
Previously we checked create-size was aligned to 512, then later we
divide it by VIXDISKLIB_SECTOR_SIZE to get the number of sectors.
VIXDISKLIB_SECTOR_SIZE is always likely to be 512, but be cleaner
about how we do this.  Also use the IS_ALIGNED macro instead of open
coding the test.

Fixes: commit a39d5773afc3ebab7e5768118a2bccb89a654585