ip: Do late filtering in list_exports as well as open
When a client lists NBD exports (eg. nbdinfo --list) then they can get
to the .list_exports callback before .open.  This could mean that a
banned client can list exports.

This is only the case when late filtering rules are in play (see
nbdkit-ip-filter(3)) which is not the normal way this filter is used.

In fact this is the case:

  $ mkdir /var/tmp/p
  $ touch /var/tmp/p/secret_export
  $ nbdkit --filter=ip file dir=/var/tmp/p allow=dn:123 deny=all \
      --run 'nbdsh --opt-mode -u "$uri" -c "h.opt_list(lambda *args: print(*args))"'
  secret_export

Above I used the 'dn:' rule, which uses late filtering.  The client
should not be able to connect at all, but can still list the exports,
thus seeing that there is a 'secret_export' (although the client
cannot open it).

When this patch is applied, the client cannot list any exports:

  $ nbdkit --filter=ip file dir=/var/tmp/p allow=dn:123 deny=all \
      --run 'nbdsh --opt-mode -u "$uri" -c "h.opt_list(lambda *args: print(*args))"'
  nbdkit: file[1]: error: client not permitted to connect because of source address restriction
  nbdsh: command line script failed: nbd_opt_list: server replied with error to list request: Operation not supported

Fixes: commit 017a4b06b48b11260197b3bc878219c38591bdf5