tests: old-plugins: Discrete tests and test using nbdinfo
Split out the monolithic old-plugins test so that each part of the
test is run separately by automake.  Also add a simple read-only test
that uses nbdinfo.

This allows us to run the test on platforms where we don't have
libguestfs, in particular on i686 which received no testing at all
previously.