tests: old-plugins: Use self-built 1.38.4 plugin instead of Fedora
On some platforms¹, testing this plugin fails with:

  + nbdkit ./old-plugins/x86_64/Linux/v1.38.4/nbdkit-file-plugin.so --version
  nbdkit: error: cannot open plugin "./old-plugins/x86_64/Linux/v1.38.4/nbdkit-file-plugin.so": /lib64/libc.so.6: version `GLIBC_ABI_DT_RELR' not found (required by ./old-plugins/x86_64/Linux/v1.38.4/nbdkit-file-plugin.so)

The RELR (relative relocation) format is explained by MaskRay here:

  https://maskray.me/blog/2021-10-31-relative-relocations-and-relr

The previous plugin was copied from a Fedora build which enables RELR,
making it incompatible with glibc on these platforms.

This commit replaces it with a plugin built by hand on RHEL 8.  It
does not use any Fedora or RHEL build flags.

As far as I know this issue should not affect i686.

¹ Platforms where CI failed:

  * AlmaLinux 9
  * CentOS Stream 9
  * Debian 11
  * OpenSUSE LEAP 15
  * Ubuntu 22.04

Fixes: commit d58218ee20ae55b90e7ee96dbfdb31a3cf6fda22