Introduce lzip-filter
This implements a filter for random access into lzip-compressed files
with multiple members. Compressed files with a single member are only
partially supported. The decompression is done using liblzma.

The function lzma_lzip_decoder() is not available on all platforms, so
there is a corresponding check in configure.ac. Unsupported platforms
include the following:

* x86_64-almalinux-9
* x86_64-centos-stream-9
* x86_64-debian-11
* x86_64-opensuse-leap-15
* x86_64-ubuntu-2204

This filter is based on the xz filter and uses filters/xz/blkcache.c.

Closes: https://gitlab.com/nbdkit/nbdkit/-/issues/23
Closes: https://gitlab.com/nbdkit/nbdkit/-/merge_requests/62
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>