Add tests for the lzip filter
This adds the following tests:

* test-lzip.c
* test-lzip-curl.c
* test-tar-lzip-curl.c

These tests are based on the corresponding tests for the xz filter and
depend on libguestfs.

* test-lzip-aligned.sh
* test-lzip-unaligned.sh

These tests ensure that the lzip filter supports compressed files with
fixed-size data blocks and variable-sized data blocks respectively.

As a prerequisite for the libguestfs-based tests, the following files are
generated by the test harness:

* disk.lz
* disk.tar.lz

All tests require lzip to be installed to generate their input data.
When lzip is missing, these tests are skipped.

The helper script chunked-lzip.sh is used to generate multi-member lzip
files with a fixed data block size, like plzip would.

Reviewed-by: Richard W.M. Jones <rjones@redhat.com>