tests: bzip2, curl, ext2, file, gzip, memory, offset, xz: Don't leak 'data' returned by guestfs_cat
This should be freed by the caller and was not.  I'm not quite certain
how this passed valgrind previously.

This error goes all the way back to commit 1b79e8ed2f ("tests: Add a
test of the gzip plugin.") in 2013.

Fixes: commit 1b79e8ed2f01c3422fc6311b252ca9d4e765ab41

Cherry picked from commit acbdc58d8046f3252177942d7609c598de2dd87b,
and remove reference to 'lzip' in the title.