ondemand: Delete the disk if creation fails
Previously if the mkfs command failed but left the disk around, then
the next time a client connected (to the same export) the incomplete
disk would still be there.  The second connection would proceed
without error and without recreating the disk correctly:

  $ rm -f /var/tmp/p/*
  $ nbdkit ondemand dir=/var/tmp/p size=10M \
           command='truncate -s $size $disk; exit 1' \
           --run 'nbdinfo --size "$uri"; nbdinfo --size "$uri"'
  nbdkit: ondemand[1]: error: command exited with code 1
  nbdinfo: nbd_get_size: server has not returned export size, you need to connect to the server first: Invalid argument
  10485760            <--- note the second connection succeeds

Fix this by deleting the incomplete disk on this error path.

Fixes: commit 0babc1ad4013f7eb4f8fbee3f41c50d0b7c414c9