tests: Use 'requires.c' code from libnbd to test requirements
In existing C tests we have code patterns such as:

  if (system ("run some command") != 0) {
    printf ("test skipped ...\n");
    exit (77);
  }

This tests for some environmental requirement or skips the test.

In libnbd we moved all this into a mini-library, so you can instead do
this (and variations):

  requires ("run some command");

Use the code from libnbd in nbdkit.  As the author, I relicensed the
libnbd code to the more permissive nbdkit license.