tcl: Fix for Tcl 9.0 compatibility
Tcl 9.0 added a new Tcl_Size typedef which is an alias for ptrdiff_t
(long int on Linux/x86_64).  Since we must pass a pointer to this type
here, get the type correct.  This uses TCL_SIZE_MAX as a sentinel
since that was also added in Tcl 9 and is not present in Tcl 8.