curl: Remove unnecessary checks for CURL_SSLVERSION_MAX_* macros
In the curl plugin we're not very consistent for how we check for
features, using one of these approaches at random:

(1) Use a ./configure-time check.  We should only do this for enums
    and functions.

(2) Use a curl version check, like LIBCURL_VERSION_MAJOR or
    CURL_AT_LEAST_VERSION.

(3) Check if a macro is defined using #ifdef.

(4) Don't check and hope the feature is available.

We were using method (1) to check for CURL_SSLVERSION_MAX_* macros,
but as these are cpp macros we can use method (3) instead for no extra
code.

We'll probably need to rationalise these checks in future and choose
one way to do it.