vddk: do_extents: Avoid reading partial chunk beyond the end of the disk
The QueryAllocatedBlocks API has (another) frustrating feature.  It
can only query whole "chunks" (128 sectors).  If the disk size is not
aligned to the chunk size (say the size was 129 sectors) then there's
a bit at the end which cannot be queried.  Furthermore, the API gives
an error in this case instead of being helpful:

  VixDiskLib_QueryAllocatedBlocks: One of the parameters was invalid

Fixes: https://issues.redhat.com/browse/RHEL-71694
Reported-by: Ming Xie <mxie@redhat.com>
Acked-by: Eric Blake <eblake@redhat.com>