curl: Fix configure-time checks for enums
On platforms such as FreeBSD 13 where curl.h is not installed on the
usual compiler search path, we NEED to include the CURL_CFLAGS as
determined by pkg-config before doing any compile-time checks for
declarations; if we don't, the compile-time check will fail because
the header wasn't found, rather than because the symbol was present or
absent; with delayed symptoms of failing to build due to attempts to
use deprecated interfaces that we would have avoided had the
declaration checks worked.  (We were already doing the similar dance
for old_LIBS later on in the AC_CHECK_FUNCS blurb; and a quick audit
of other AC_CHECK_DECL and AC_CHECK_FUNCS after PKG_CHECK_MODULES
didn't find any other offenders).

Fixes: 9ed65418 ("curl: Remove unnecessary checks for CURL_SSLVERSION_MAX_* macros", v1.41.9)
Fixes: 9408b8c7 ("curl: Use CURLOPT_PROTOCOLS_STR if available", v1.33.8)
Signed-off-by: Eric Blake <eblake@redhat.com>