server: Allow user statically defined tracing (USDT) probes
This commit allows USDT probes to be freely added to the server.  (But
not to plugins without some extra work.)  In this commit only two
actual probes have been added.  The idea is that developers would add
more probes as required.

Note the overhead of the probes, when disabled, is negligible even on
hot paths.  It basically just inserts a NOP.  When enabled, the NOP
instruction is replaced with a trap into the kernel (eg. INT3 on x86)
which can be traced using perf or other tools.

Some useful documents covering these probes:
https://sourceware.org/systemtap/wiki/AddingUserSpaceProbingToApps
https://blog.vmsplice.net/2017/07/tracing-userspace-static-probes-with.html
https://jvns.ca/blog/2017/07/05/linux-tracing-systems/
https://docs.openvswitch.org/en/latest/topics/usdt-probes/

This large change also includes documentation and enables CI.