rust: Disable clippy warnings unless --enable-gcc-warnings is used
Turn ./configure --enable-gcc-warnings into a general "enable compiler
warnings" option (we'll probably need to change the name of it later).
Then use this option to control whether Rust clippy warnings are
enabled or not.

Note that CI generally enables this option so we should still see Rust
errors in CI.

Reported-by: Thomas Weißschuh