protostrings: Use correct path to nbd-protocol.h
In RHEL, nbdkit didn't print things like protocol commands properly.
For example it would print:

  nbdkit: null.0: debug: client sent unknown (0x2), closing connection

instead of:

  nbdkit: null.0: debug: client sent NBD_CMD_DISC, closing connection

Building from source worked.

This happened because the script that generates the protostrings code
doesn't use the correct path to nbd-protocol.h relative to the source
directory.  In RHEL we use a separate build subdirectory, so the
script failed to read nbd-protocol.h (silently, because it's a shell
script) and misgenerated the printing functions.