common/protocol: Use thread-local storage for "unknown" string
If we pass an unknown value to functions like name_of_nbd_cmd, it
returns a string like "unknown (0x100)".  However this string was
globally allocated, so different threads would conflict with each
other.  This in itself doesn't seem like anything more than a cosmetic
function - after all, these strings are only used in the supposedly
rare case where an unknown value is seen, not on the happy path.  The
only bad thing that could happen is garbled output in debug messages.

However:

(1) The parameter values come from untrusted user input, meaning this
is not necessarily rare, and could be controlled by an attacker.

(2) Because of an unrelated bug (fixed in commit 86dd0c76b0) the
unhappy path was happening all the time in some downstream
distributions like RHEL.

(3) The debug messages are useful for debugging, so we ought to
get them right.

(4) The functions wouldn't work if we call the same function twice in
a single statement (eg. from printf).  Actually this doesn't happen in
existing code.

The fix is relatively easy.  As it's a statically allocated buffer we
can just put it in thread-local storage, avoiding problems (1)-(3).

For problem (4) I checked the existing code.

(cherry picked from commit f190ad70bcba8540d5a6e74488ad388cc8bc1a62)