file: If sync_file_range fails to start, don't add window to list
When starting the process of writing out the newest range, we
previously didn't check if sync_file_range succeeded or failed.  But
if it failed there's no reason to add it to the list of windows.  Also
debug any failures.

This shouldn't change any semantics as far as we're aware.

Thanks: Eric Sandeen