tests: Add tests of file plugin cache=none
These are based on the tests which were done in the commit message
when the feature was first added in commit aa5a2183a6, but cleaned up
and modernised a little.

It is likely that these tests would fail under memory pressure (since
they test that the page cache is used vs not used).  Making them more
robust left as an exercise for the future.