New plugin: NFS plugin using libnfs
This requires libnfs >= 6 (internal version 16).

This is missing some fairly vital support for discard and extents, so
performance is not that great.  On non-sparse files it is reasonably
fast, but it's much slower on sparse files because it has to read
zeroes over the wire.

This works well against the Linux and Synology NFS servers (presumably
the same thing).

It doesn't work against unfs3 reporting the libnfs error "Fragment
support not yet working".  This is a shame because a working userspace
NFS server would allow us to implement tests.  As it is there are no
tests included.

There are no tests as this is very difficult to test without needing
root and system-level changes.

Thanks: Ronnie Sahlberg
See-also: https://groups.google.com/g/libnfs/c/Mq0Ialk2k2o