tests/test-golang-fork-warning.sh: Don't use nbdinfo in test
This test hangs in gitlab CI.  Although I don't have direct evidence,
a possible explanation is that after fork, the golang plugin really is
broken.  (This is, after all, the purpose of having a warning — some
background thread needed by golang to service requests is missing.)
As a result when nbdinfo calls into nbdkit it causes a hang.

Whatever the true reason, we can simplify the test by removing the
unnecessary use of nbdinfo.

Thanks: Peter Krempa
Fixes: commit 2589e6da40939af9aea16a488e8f2aec411f3c5c