tests: More robust check for nbdsh --base-allocation
All tests that rely on 'nbdsh --base-allocation' working (basically
because they are testing .extents behavior) were using the same
copy-paste 'requires nbdsh --base-allocation'.  This happens to work
because requires redirects stdin from /dev/null, but is a bit awkward
because it would go interactive otherwise.  Pairing with --verbose to
force non-interactive even if stdin is not redirected is better.

Suggested-by: Rich Jones <rjones@redhat.com>
Signed-off-by: Eric Blake <eblake@redhat.com>