file: Fix check for errno EOPNOTSUPP
Commit abb2b47c8ad ("plugins: Permit ENOTSUP as synonym for
EOPNOTSUPP") changed several errno tests using this general (and
correct) pattern:

-    if (errno != EOPNOTSUPP) {
+    if (!is_enotsup (errno)) {

However in one place it was changed to the nonsensical
'!is_enotsup (EOPNOTSUPP)' by accident.  Fix this.

The error is not very serious.  It means that we won't notice when
FALLOC_FL_PUNCH_HOLE is unsupported by the kernel, and keep retrying
it.

The same error was apparently copied to the ondemand and tmpdisk
plugins, so fix it there too.

Thanks: Eric Blake
Fixes: commit abb2b47c8adfca63ca81f46b038d1a57a8662998