server: Move saved exportname from connection to context
The exportname was stored in the connection because, in normal
operation, there's only one connection, but there's a context for each
filter and the plugin, so storing the exportname just once made sense.

However for shared contexts, there is no connection (context->conn is
NULL).  Thus there is no place to store the exportname, and no way for
nbdkit_export_name to get it (it returned NULL).  But shared contexts
are created with an exportname parameter (to nbdkit_next_context_open)
so returning NULL from nbdkit_export_name is wrong.

Moving the exportname field to the context structure solves the
problem that nbdkit_export_name couldn't get the exportname even
though it exists, at the expense of storing the exportname multiple
times, once in each layer.

I also made the exportname a standard allocated string with the
lifetime the same as the context object.  This is because shared
contexts have an uncertain lifetime, but interning would probably
choose the connection lifetime which could be too short.