ci: Install python3-packaging for the CI tests
Tests which use the requires_libnbd_version function need python
"packaging" (a module), as we use that to detect the version of libnbd
and compare versions.  This means that any tests using that function
aren't actually running presently.  Fix that by installing
python3-packaging or its equivalents on different distros.

This requires the following change to libvirt-ci:
https://gitlab.com/libvirt/libvirt-ci/-/merge_requests/534