vddk: Unconditionally test QueryAllocatedBlocks
Except in rare cases like a suddenly dropped connection, we always
call vddk_can_extents and therefore do this test.  We might as well do
it unconditionally when the worker thread starts up.

This simplifies following commits where we may do more work based on
this flag when the worker thread starts up.