server: Align the truncation of large block_status results
The NBD spec currently requires that block status results be aligned
to any advertised minimum block size (although I'm submitting a patch
in parallel to relax that slightly in the face of unaligned file size,
to cater to the fact that not all existing servers comply, such as
nbdkit 1.42).  And while most plugins and filters _do_ happen to
provide aligned results when given aligned inputs (such as with the
previous patch to blocksize-policy), we have a corner case where an
aligned request that results in a 4G+ aligned response from the plugin
is truncated in the server code to an unaligned answer.

For background, prior to advertising block sizes, the server didn't
have to care about alignments: the spec only requires aligned results
when the client asks and the server replies with a minimum block size.
And even with the introduction of the .block_size callback, the server
still doesn't have to care for reads or writes: those responses will
naturally have the same alignment as the request (theoretically, we'd
have to worry about alignment for reads split into chunks when taking
advantage of reporting holes efficiently via NBD_CMD_READ - but we
don't support split reads until we design the v3 plugin API).  It is
_only_ for extents where even if the client passes an aligned request
in, we are at the mercy of whether the plugin reports extents smaller
than the alignment.  But we have one exception: even when the plugin
dutifully guarantees aligned block extents, if it widened the user's
aligned request for a block status less than 4G into an aligned value
4G or larger, then the plugin no longer has control of how the server
further splits that response into something that will fit in the
32-bit response over the wire to the client.  If the server misbehaves
there, it WILL confuse clients such as 'nbdinfo --map', which fails in
libnbd 1.22 when it later tries to call an unaligned block status to
learn about the rest of the image.  Libnbd 1.24 has added a patch to
work around this slop near 4G, but this patch allows even older libnbd
to fully map images that contain aligned extents larger 4G.

The fix here is done by observing that .block_size cannot set a
minimum block size larger than 64k.  Instead of having to teach the
server code about alignment everywhere (reads, writes, all aspects of
block status), it is far easier to just truncate the server's response
at a point that is guaranteed to be aligned for all possible plugins,
and continue to leave plugins in charge of alignment.  By default,
nbdkit still handles unaligned requests just fine, and the
blocksize-policy filter is enough to enforce the alignment when
enforcement is wanted.

The change here is observable in test-eval-extents.sh which was added
as part of the fix for CVE-2025-47711; but fortunately, spec
non-compliance is different than denial of service, so versions of
nbdkit with this bug do NOT represent yet another security lapse, and
the changed test is not any weaker.

Signed-off-by: Eric Blake <eblake@redhat.com>