blocksize-policy: Round extents to match minimum alignment
Since a client is very likely to make further I/O calls at offsets
returned in a block status call (for example, 'nbdinfo --map
--totals'), and the blocksize-policy filter can set a minimum
alignment larger than what the plugin does by default, it makes sense
to force extents to that alignment.  In particular, this problem with
unaligned extents was happening even when the image size is aligned,
and violates the NBD spec that says that block status results should
be aligned to the minimum block size when both client and server have
agreed to a minimum; and even if you use newer libnbd 1.24 that works
around misalignment near 4G, this one is still something better fixed
in the server rather than the client.

Signed-off-by: Eric Blake <eblake@redhat.com>