data: Allow scaling suffixes on @decimal
Although you can write 'nbdkit data "@$((4*1024)) 1"', it is more
convenient to be able to write 'nbdkit data "@4k 1"'.  To do that,
make use of our existing code for suffix scaling when we know we have
a decimal value, while still accepting non-decimal values without
scaling for backwards compatability.  Since we must do substring
parsing (there is likely more data after the offset directive), we
insist that after the value is parsed, the string must be on
whitespace or closing ')' before the next directive.

Signed-off-by: Eric Blake <eblake@redhat.com>