human-size: Add human_size_parse_substr
I want to teach the data plugin how to parse @4k out of a larger input
string, but that entails the need for a way to opt in to not erroring
out when an unrecognized suffix is present.

Some of the existing unit tests expose cases where parsing a value
substring out of a larger string can have interesting results; but in
the case of the data plugin, the caller will validate that any
remaining suffix starts with whitespace (and not '.' or a letter),
which should avoid the oddest cases due to the additional context that
the unit test lacks.

While testing this, I ran into a gcc false positive[1], which in turn
was more readily triggered by a glibc weakness in the signature of
strtoimax[2].
[1] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=120526
[2] https://sourceware.org/bugzilla/show_bug.cgi?id=33053

Signed-off-by: Eric Blake <eblake@redhat.com>