blocksize-policy: Document use with truncate filter
Current libnbd defaults to insisting that alignment be honored, even
when the advertised size is unaligned.  While I am also working on
patches to make 'nbdinfo --map' be a little more aggressive in trying
unaligned access at the tail of an image, it is worth documenting that
the truncate filter is useful to pair with the blocksize-policy filter
any time minimum block alignment is larger than one.

Signed-off-by: Eric Blake <eblake@redhat.com>