extents: Handle unaligned tail in nbdkit_extents_aligned
All existing callers of nbdkit_extents_aligned (namely: blocksize,
swab) truncate the image size to an aligned value, so asserting that
count is aligned was okay.  But an upcoming patch will change the
blocksize-policy filter to ask for aligned data while still permitting
access to the unaligned tail of an image.  As such, it is time to
handle that special case by treating the unaligned tail of an image as
if it had a smaller alignment constraint of that tail in isolation.

Signed-off-by: Eric Blake <eblake@redhat.com>