blocksize-policy: Always flag unaligned block_status calls
If the plugin's .can_extents reports 0, then the server's backend code
was synthesizing a block status result of all data without calling
into the filter, regardless of alignment.  Although this is fine for
blocksize-error-policy=allow, it meant that the behavior for
blocksize-error-policy=error was unable to flag unaligned block_status
calls.

Fixes: 1747ef0f ("New filter: nbdkit-blocksize-policy-filter", v1.29.16)
Signed-off-by: Eric Blake <eblake@redhat.com>