maint: Better fix for clang
The recent patch to try and fix the build on fedora + clang++ missed
the fact that on FreeBSD, the compiler is named 'c++' rather than
'clang++'.  I'm not sure why ax_pthread.m4 isn't getting everything to
work right for c++ since it is already special-casing -pthread when cc
is clang, but this hack is simple enough, and more robust than
checking a compiler's name.  But it does depend on running AX_PTHREAD
before we mess with CXXFLAGS.

Fixes: 2e557ba9 ("maint: Silence clang warning")
Signed-off-by: Eric Blake <eblake@redhat.com>
(cherry picked from commit d417cdc08909559678cfc92158b3609a9bc1c594)