vddk: stats: Collect elapsed time for ReadAsync and WriteAsync
A major existing problem with -D vddk.stats=1 is that it did not
accurately account for the time taken in these asynchronous calls.
This change records the elapsed time.  (It's not possible to collect
the actual time spent "working on" these calls, we just add up the
total elapsed time, which is fine.)

In the remote case, here's how the output changes from before this
change:

nbdkit: debug: VixDiskLib_...                    us calls           bytes
nbdkit: debug:   ReadAsync                       18     1            8192

to after this change:

nbdkit: debug: VixDiskLib_...                    us calls           bytes
nbdkit: debug:   ReadAsync                   243745     1            8192

An interesting thing about this is that in the local case this
actually reduces the apparent elapsed time, I think because VDDK calls
complete_command directly from the Read/WriteAsync function.  It only
increases the apparent elapsed time in the remote case.